Rename docker modules
